\name{CARLIS-package}
\alias{CARLIS-package}
\alias{CARLIS}
\docType{package}
\title{
\packageTitle{CARLIS}
}
\description{
\packageDescription{CARLIS}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{CARLIS}

Maintainer: \packageMaintainer{CARLIS}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = CARLIS)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{CARLIS}}
%\section{Documentation Index}{\packageIndices{CARLIS}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
